+++ src/main/java/de/dealog/msg/persistence/repository/MessageRepository.java
@@ -1,7 +1,7 @@
 package de.dealog.msg.persistence.repository;
 
 import de.dealog.msg.persistence.model.MessageEntity;
-import de.dealog.msg.persistence.model.MessageStatus;
+import de.dealog.common.model.Status;
 import io.quarkus.hibernate.orm.panache.PanacheRepository;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -13,7 +13,7 @@ public class MessageRepository implements PanacheRepository<MessageEntity> {
         return find("identifier", identifier).firstResult();
     }
 
-    public MessageEntity findByIdentifierAndStatus(final String identifier, final MessageStatus status){
+    public MessageEntity findByIdentifierAndStatus(final String identifier, final Status status){
         return find("identifier = ?1 and status = ?2", identifier, status).firstResult();
     }
 }