+++ src/main/java/de/dealog/msg/rest/MessageConverter.java
@@ -14,9 +14,11 @@ public class MessageConverter extends Converter<Message, MessageRest> {
     protected MessageRest doForward(final Message message) {
         final MessageRest messageRest= new MessageRest();
         messageRest.setIdentifier(message.getIdentifier());
+        messageRest.setOrganization(message.getOrganization());
         messageRest.setHeadline(message.getHeadline());
         messageRest.setDescription(message.getDescription());
         messageRest.setArs(message.getRegionCode());
+        messageRest.setCategory(message.getCategory());
         messageRest.setPublishedAt(message.getPublishedAt());
         return messageRest;
     }
@@ -25,9 +27,11 @@ public class MessageConverter extends Converter<Message, MessageRest> {
     protected Message doBackward(final MessageRest messageRest) {
         final MessageEntity message= new MessageEntity();
         message.setIdentifier(messageRest.getIdentifier());
+        message.setOrganization(messageRest.getOrganization());
         message.setHeadline(messageRest.getHeadline());
         message.setDescription(messageRest.getDescription());
         message.setRegionCode(messageRest.getArs());
+        message.setCategory(messageRest.getCategory());
         message.setPublishedAt(messageRest.getPublishedAt());
         return message;
     }